Changed DummyLogger in favor of ArrayLogger
