+++ tests/Elasticsearch/Tests/ClientBuilder/DummyLogger.php
@@ -3,55 +3,7 @@ declare(strict_types = 1);
 
 namespace Elasticsearch\Tests\ClientBuilder;
 
-use Psr\Log\LoggerInterface;
-use Psr\Log\LogLevel;
-
-class DummyLogger implements LoggerInterface
+class DummyLogger
 {
-    public $output = [];
-
-    public function emergency($message, array $context = array())
-    {
-        $this->log(LogLevel::EMERGENCY, $message, $context);
-    }
-
-    public function alert($message, array $context = array())
-    {
-        $this->log(LogLevel::ALERT, $message, $context);
-    }
-
-    public function critical($message, array $context = array())
-    {
-        $this->log(LogLevel::CRITICAL, $message, $context);
-    }
-
-    public function error($message, array $context = array())
-    {
-        $this->log(LogLevel::ERROR, $message, $context);
-    }
-
-    public function warning($message, array $context = array())
-    {
-        $this->log(LogLevel::WARNING, $message, $context);
-    }
-
-    public function notice($message, array $context = array())
-    {
-        $this->log(LogLevel::NOTICE, $message, $context);
-    }
-
-    public function info($message, array $context = array())
-    {
-        $this->log(LogLevel::INFO, $message, $context);
-    }
-
-    public function debug($message, array $context = array())
-    {
-        $this->log(LogLevel::DEBUG, $message, $context);
-    }
-
-    public function log($level, $message, array $context = array())
-    {
-        $this->output[] = sprintf("%s: %s %s", $level, $message, json_encode($context));
-    }
+    
 }
